+++ plumage/static/js/main.js
@@ -1,23 +1,22 @@
-
 // Activate Bootstrap's tooltips
 const tooltipTriggerList = document.querySelectorAll('[data-bs-toggle="tooltip"]')
-const tooltipList = [...tooltipTriggerList].map(tooltipTriggerEl => new bootstrap.Tooltip(tooltipTriggerEl))
+tooltipTriggerList.forEach(tooltipTriggerEl => new bootstrap.Tooltip(tooltipTriggerEl))
 
 // Fix navigation to the top on scroll.
-var nav = document.querySelector(".navbar")
-var brand = document.querySelector(".navbar-brand")
-var navTop = nav.offsetTop
-var isFixed = 0
-window.onscroll = function (event) {
+const nav = document.querySelector('.navbar')
+const brand = document.querySelector('.navbar-brand')
+const navTop = nav.offsetTop
+let isFixed = false
+window.onscroll = function () {
   if (window.scrollY >= navTop && !isFixed) {
-    isFixed = 1
-    nav.classList.add('fixed-top');
-    nav.classList.remove("rounded-3", "d-block");
-    brand.classList.remove('d-none');
+    isFixed = true
+    nav.classList.add('fixed-top')
+    nav.classList.remove('rounded-3', 'd-block')
+    brand.classList.remove('d-none')
   } else if (window.scrollY <= navTop && isFixed) {
-    isFixed = 0
-    nav.classList.remove('fixed-top');
-    nav.classList.add("rounded-3", "d-block");
-    brand.classList.add('d-none');
+    isFixed = false
+    nav.classList.remove('fixed-top')
+    nav.classList.add('rounded-3', 'd-block')
+    brand.classList.add('d-none')
   }
-};
+}